Filter files by size Filter latest files by time span Launcher automatically selecting 32/64 bit executable on start up More detailed systray progress indicator New database format for <Automatic> mode: a full sync is suggested before upgrading Update database at individual file level (support for partial and aborted syncs) New translation file format Dynamically load existing translation files Correct translation plural forms Improved directory locking strategy Restructured installation package One button-click synchronization Fixed CSV character encoding Put CSV values in quotes if they contain semicolons Explicit button and settings for "Custom" sync variant -> old configurations need to be migrated Keyboard shortcuts also on middle grid Minimize progress dialog by clicking on taskbar Render invalid file dates correctly on GUI Process user-defined commands via shell execution (FFS and RTS) Allow base directory names having trailing white-space Added Ukrainian language Updated translation files

+++ ui/tray_icon.h
@@ -17,7 +17,7 @@ public:
     MinimizeToTray(wxTopLevelWindow* callerWnd, wxTopLevelWindow* secondWnd = NULL); //ensure both windows have longer lifetime than this instance!
     ~MinimizeToTray(); //show windows again
 
-    void setToolTip(const wxString& toolTipText, size_t percent = 0); //percent (optional), number between [0, 100], for small progress indicator
+    void setToolTip(const wxString& toolTipText, double percent = 0); //percent (optional), number between [0, 100], for small progress indicator
     void keepHidden(); //do not show windows again: avoid window flashing shortly before it is destroyed
 
 private:
